Refine XPath evaluator split
